+++ core/model/src/main/java/com/example/model/Goods.kt
@@ -3,9 +3,9 @@ package com.example.model
 data class Goods(
     val id: Long,
     val name: String,
-    val price: Int,
-    val purchases: Int,
-    val remain: Int,
+    val price: Long,
+    val purchases: Long,
+    val remain: Long,
     val isAvailable: Boolean,
     val displayOrder: Long,
 )
